chore: apply clippy suggestions for paths, closures, and casts
- Change `&PathBuf` to `&Path` in `migrate` and `rename` signatures to
  avoid unnecessary allocations.
- Simplify `io::Error::new(io::ErrorKind::Other, ...)` to
  `io::Error::other`.
- Remove redundant closures in iterator chains and lazy unwraps.
- Remove unnecessary `u64` type casting on `DEFAULT_STEP`.
- Remove unused `dirs` import from `src/file.rs`.

+++ src/otp.rs
@@ -20,7 +20,7 @@ fn get_current_timestamp() -> Result<u64, OtpError> {
 pub fn get_remaining_seconds() -> u64 {
     let now = get_current_timestamp().unwrap_or(0);
     // DEFAULT_STEP == 30s
-    DEFAULT_STEP as u64 - (now % DEFAULT_STEP as u64)
+    DEFAULT_STEP - (now % DEFAULT_STEP)
 }
 
 /*
